refactor current bootloader reading out of the server ino ota
